Remove connection pooling.
The connection pooling logic in this code adds a certain amount
of fragility, as the connections can wind up being reused even
after the remote service has restarted.

Ultimately, this service needs to be rethought so that it detects
a remote connection close, and discards connections that have
aborted.  This would require adding a pending read against the
connection, in order to get notification that the remote socket
has completed.

Unfortunately, the current design is too simplistic to easily add
that.  Removing the connection pooling may at least alleviate some
pain, as a restart will only affect transactions that are in flight,
at that time, but if the connection start up time is high (e.g. setting
up a TLS protected TCP channel), then performance will drop.

On the plus side, the removal of the pool means that clients will
potentially achieve much greater levels of parallelization, as
the pool size no longer limits this.  This is useful when running
many requests that may be parallel but take a long time (such as
requests that poll for physical disks.)